feat: logging and new entities

+++ Homer.NetDaemon/codegen.ps1
@@ -5,4 +5,6 @@ Write-Output "Using secret: $secret"
 
 dotnet tool run nd-codegen -o "Entities/Entities.cs" -ns "Homer.NetDaemon.Entities" -token $secret
 
-Remove-Item -Recurse NetDaemonCodeGen
+if ($?) {
+    Remove-Item -Recurse NetDaemonCodeGen
+}
